new participant list component now linked to current user page
